Make module private
