+++ src/aiida/cmdline/commands/cmd_process.py
@@ -629,8 +629,8 @@ def process_dump(
 
     from aiida.cmdline.utils import echo
     from aiida.tools.archive.exceptions import ExportValidationError
-    from aiida.tools.dumping.config import DumpConfig, DumpMode
-    from aiida.tools.dumping.utils import DumpPaths
+    from aiida.tools._dumping.config import DumpConfig, DumpMode
+    from aiida.tools._dumping.utils import DumpPaths
 
     warning_msg = (
         'This is a new feature which is still in its testing phase. '
