+++ src/aiida/cmdline/commands/cmd_profile.py
@@ -356,8 +356,8 @@ def profile_dump(
     from pydantic import ValidationError
 
     from aiida.cmdline.utils import echo
-    from aiida.tools.dumping.config import DumpConfig, DumpMode
-    from aiida.tools.dumping.utils import DumpPaths
+    from aiida.tools._dumping.config import DumpConfig, DumpMode
+    from aiida.tools._dumping.utils import DumpPaths
 
     warning_msg = (
         'This is a new feature which is still in its testing phase. '
