--- src/aiida/tools/dumping/utils/__init__.py
@@ -1,42 +1,0 @@
-###########################################################################
-# Copyright (c), The AiiDA team. All rights reserved.                     #
-# This file is part of the AiiDA code.                                    #
-#                                                                         #
-# The code is hosted on GitHub at https://github.com/aiidateam/aiida-core #
-# For further information on the license, see the LICENSE.txt file        #
-# For further information please visit http://www.aiida.net               #
-###########################################################################
-"""Utilities for the dumping feature."""
-
-from aiida.tools.dumping.utils.helpers import (
-    ORM_TYPE_TO_REGISTRY,
-    REGISTRY_TO_ORM_TYPE,
-    DumpChanges,
-    DumpTimes,
-    GroupChanges,
-    GroupInfo,
-    GroupModificationInfo,
-    GroupRenameInfo,
-    NodeChanges,
-    NodeMembershipChange,
-    ProcessingQueue,
-    RegistryNameType,
-)
-from aiida.tools.dumping.utils.paths import DumpPaths
-
-__all__ = (
-    'ORM_TYPE_TO_REGISTRY',
-    'REGISTRY_TO_ORM_TYPE',
-    'DumpChanges',
-    'DumpPaths',
-    'DumpTimes',
-    'DumpTimes',
-    'GroupChanges',
-    'GroupInfo',
-    'GroupModificationInfo',
-    'GroupRenameInfo',
-    'NodeChanges',
-    'NodeMembershipChange',
-    'ProcessingQueue',
-    'RegistryNameType',
-)
